requirements.txt
bcrypt==3.1.4
cffi==1.11.2
Django==2.0.1
django-filter==1.1.0
djangorestframework==3.7.7
git+https://github.com/soynatan/django-easy-audit@3518a7b3d6225f6b0340c920ebedfad9490eafe1
django-session-security==2.6.0
gunicorn==19.7.1
psycopg2==2.7.3.2
pycparser==2.18
pytz==2017.3
six==1.11.0
arrow==0.12.0
django-health-check==3.3.0
django-fernet-fields==0.5
django-auth-ldap==1.5.0
python-ldap==3.0.0
